refactor(block-editor): extract useCodaOptions to a shared hook
Pulls the hook out of TerminalConnectBlockForm where it was defined
inline. ChallengeBlockForm will reuse it in the next commit so its
App-name and Scenario fields can become real comboboxes against the
Coda catalog instead of free-text inputs.

No behaviour change for the terminal-connect block.
